Ajuste de id, int -> Long

+++ CrudPersonaBackEnd/src/main/java/com/example/demo/model/Pais.java
@@ -7,6 +7,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -19,7 +20,7 @@ public class Pais {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "pai_id")
-	private int id;
+	private Long id;
 	
 	@Column(name = "pai_nombre")
 	private String nombre;
